[update] firebase apn 관련 설정 sceneDelegate로 옮김
