+++ B-umiOS/B-umiOS/Resource/Supporting Files/SceneDelegate.swift
@@ -5,18 +5,20 @@
 //  Created by inae Lee on 2021/06/29.
 //
 
+import Firebase
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
     var window: UIWindow?
-
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let _ = (scene as? UIWindowScene) else { return }
+
+        UNUserNotificationCenter.current().delegate = self
+        Messaging.messaging().delegate = self
     }
 
     func sceneDidDisconnect(_ scene: UIScene) {
@@ -46,7 +48,56 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use this method to save data, release shared resources, and store enough scene-specific state information
         // to restore the scene back to its current state.
     }
-
-
 }
 
+extension SceneDelegate: UNUserNotificationCenterDelegate {
+    /// foreground 상태일 때 push 알림 받음
+    func userNotificationCenter(_ center: UNUserNotificationCenter, willPresent notification: UNNotification, withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
+        let userInfo = notification.request.content.userInfo
+        print(userInfo, "👅 FOREGROUND!")
+
+        /// 애널리틱스에 전달
+        Messaging.messaging().appDidReceiveMessage(userInfo)
+        completionHandler([.list, .badge, .sound, .banner])
+    }
+
+    /// background일 때
+    func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
+        UIApplication.shared.applicationIconBadgeNumber += 1
+
+        let userInfo = response.notification.request.content.userInfo
+        print(userInfo, "👅")
+
+        if let aps = userInfo["aps"] as? [String: AnyObject] {
+            print(aps)
+        }
+
+        Messaging.messaging().appDidReceiveMessage(userInfo)
+        completionHandler()
+    }
+}
+
+extension SceneDelegate: MessagingDelegate {
+    func applicationDidFinishLaunching(_ application: UIApplication) {
+        print("applicationDidFinishLaunching")
+    }
+
+    /// 현재 등록 토큰 가져오기 (갱신 시 알림 받기)
+    func messaging(_ messaging: Messaging, didReceiveRegistrationToken fcmToken: String?) {
+        print("🥺 [fcm token]: ", fcmToken)
+        guard let token = fcmToken else { return }
+
+        UserService.shared.registerFCMToken(token: fcmTokenRequest(pushToken: token))
+    }
+
+    func application(_ application: UIApplication, didReceiveRemoteNotification userInfo: [AnyHashable: Any], fetchCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void) {
+        print(userInfo, "💬")
+        Messaging.messaging().appDidReceiveMessage(userInfo)
+
+        completionHandler(UIBackgroundFetchResult.newData)
+    }
+
+    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
+        Messaging.messaging().apnsToken = deviceToken
+    }
+}
